requirements.txt
numpy==1.22.4
opencv-python==4.5.5.64
Pillow==9.1.1
typing_extensions==4.2.0
